Tests/LibWeb/Text/expected/canvas/drawimage-canvas-self.txt

right half received the pre-draw left half (blue): true
left half is untouched (blue): true
